api/requirements.txt
Flask==2.1.2
Flask_Cors==3.0.10
Flask_Login==0.6.2
Flask_SQLAlchemy==2.5.1
Flask_Session==0.4.0
PyJWT==2.6.0
Routes==2.5.1
SQLAlchemy==1.3.24
Werkzeug==2.1.2
Flask-Compress==1.13
Brotli==1.0.9
flask_restx==1.0.6
geopandas==0.10.2
gunicorn==20.1.0
passlib==1.7.4
psycopg2-binary==2.9.3
pyproj==3.3.1
pytest==7.2.1
Shapely==1.8.4
